example-microservices/oidc/requirements.txt
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Django==2.1.4
django-oidc-provider==0.7.0
future==0.17.1
idna==2.7
pycparser==2.19
pycryptodomex==3.7.2
pyjwkest==1.4.0
PyNaCl==1.3.0
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
